Rename 'cloud' folder to 'cluster' for clearer association
