+++ cluster/store/schema_test.go
