+++ cluster/store/snapshot.go
